added logic and ui to update password of users choice by sending verification link to email

+++ codingthoughts/src/app/api/update/emailVerify/route.ts
@@ -25,11 +25,6 @@ export async function POST(req: Request) {
     if(token.expiresAt < new Date()) {
         return NextResponse.json({ message: "This link has expired. Please save your email again to request a new link."}, { status: 400 });
     }
-    // check token type
-    if(token.type != "EMAIL_CHANGE") {
-        console.log("Invalid token type");
-        return NextResponse.json({ message: errorMessage}, { status: 400 });
-    }
     // valid token, check if matches to a userId
     const user: User | null = await prisma.user.findUnique({
         where: {
@@ -43,31 +38,47 @@ export async function POST(req: Request) {
             , { status: 400 }
         );
     }
-    // valid user, change this users email to the new one
-    try {
-        await prisma.user.update({
-            where: { id: user.id },
-            data: { email: token.value }
-        });
-    } catch (error) {
-        console.log("Failed to update user email", error);
-        return NextResponse.json({ message: errorMessage }, { status: 500 });
-    }
-    // delete all expired tokens
-    await prisma.token.deleteMany({
-        where: {
-            expiresAt: {
-            lt: new Date(), // Deletes tokens where expiry is in the past
-            },
-        },
-    });
-    // delete current token
-    await prisma.token.delete({
-        where: {
-            id: tokenId
+    // check token type
+    if(token.type != "EMAIL_CHANGE") {
+        // password change token, change the password and return
+        try {
+            await prisma.user.update({
+                where: { id: user.id },
+                data: { password: token.value }
+            });
+        } catch (error) {
+            console.log("Failed to update user password", error);
+            return NextResponse.json({ message: errorMessage }, { status: 500 });
         }
-    });
-    // otherwise all good
-    return NextResponse.json
-    ({ message: "Email changed successfully. You may return to your old window and see your updated account details. Please wait up to 5 seconds if your email has not changed."});
+        // delete all expired tokens
+        await prisma.token.deleteMany({
+            where: {
+                expiresAt: {
+                lt: new Date(), // Deletes tokens where expiry is in the past
+                },
+            },
+        });
+        // delete current token
+        await prisma.token.delete({
+            where: {
+                id: tokenId
+            }
+        });6
+        return NextResponse.json({ message: "Password changed successfully, you may log in to your account now."});
+    }
+    else {
+        // email change token, change the email
+        try {
+            await prisma.user.update({
+                where: { id: user.id },
+                data: { email: token.value }
+            });
+        } catch (error) {
+            console.log("Failed to update user email", error);
+            return NextResponse.json({ message: errorMessage }, { status: 500 });
+        }
+        // otherwise all good
+        return NextResponse.json
+        ({ message: "Email changed successfully. You may return to your old window and see your updated account details. Please wait up to 5 seconds if your email has not changed."});
+    }
 }
